% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_pg.R
\name{connect_to_pg}
\alias{connect_to_pg}
\title{Connect to the Postgres server}
\usage{
connect_to_pg(host = NULL, port = NULL, user = NULL, pass = NULL)
}
\arguments{
\item{host}{character, host name of pg server}

\item{port}{character, port of pg server}

\item{user}{character, username for pg server}

\item{pass}{character, password for pg server}
}
\value{
Connection object for pg queries
}
\description{
This connects to the PG server, acting as a wrapper
}
