% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dev_ecte.R
\name{create_dev_ecte}
\alias{create_dev_ecte}
\title{Create dev_ecte pg table}
\usage{
create_dev_ecte()
}
\arguments{
\item{con}{connection object}

\item{data}{data.frame/data.table port of pg server}

\item{table}{character, name of pg table}
}
\value{
NA
}
\description{
Creates dev_ecte table with proper schema, index and partitions
}
