% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ec.R
\name{download_ec}
\alias{download_ec}
\title{Wrapper to pull EC data from NEON portal and write it to a PG database}
\usage{
download_ec(
  site = "PUUM",
  start_date = "2017-01-01",
  end_date = Sys.Date(),
  package = "basic"
)
}
\arguments{
\item{site}{character, 4 letter site code}

\item{start_date}{character/date, yyyy-mm-dd, start date to pull}

\item{end_date}{character/date, yyyy-mm-dd, end date to pull}

\item{package}{character, extended or basic?}
}
\value{
NA
}
\description{
This function pulls data, extracts data from HDF5, and saves the data out to a PG server
}
