% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_collect_hdf5_data.R
\name{init_collect_hdf5_data}
\alias{init_collect_hdf5_data}
\title{Collect the data from an HDF5 file}
\usage{
init_collect_hdf5_data(path = NULL)
}
\arguments{
\item{path}{character, path to the h5 file}
}
\value{
the data in the HDF5
}
\description{
This opens the HDF5 file at the path and identifies what streams to pull, and grabs them from the file
}
