% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_msg.R
\name{init_msg}
\alias{init_msg}
\title{Wrapper to message out during init functions}
\usage{
init_msg(msg = NULL)
}
\arguments{
\item{msg}{a message, character vector}
}
\value{
A standard message '2023-02-08 20:37:48 MST: This is a sample message'
}
\description{
This takes a char vector and applies standard messaging
}
