% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into_pg.R
\name{insert_into_pg}
\alias{insert_into_pg}
\title{Insert data into the table}
\usage{
insert_into_pg(con = NULL, data = NULL, table = NULL)
}
\arguments{
\item{con}{connection object}

\item{data}{data.frame/data.table port of pg server}

\item{table}{character, name of pg table}
}
\value{
NA
}
\description{
This inserts data into the specified table
}
