% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_connect.R
\name{pg_connect}
\alias{pg_connect}
\title{Connect to the Postgres server}
\usage{
pg_connect(host = NULL, port = NULL, user = NULL, pass = NULL)
}
\arguments{
\item{host}{character, host name of pg server}

\item{port}{character, port of pg server}

\item{user}{character, username for pg server}

\item{pass}{character, password for pg server}
}
\value{
Connection object for pg queries
}
\description{
This connects to the PG server, acting as a wrapper
}
