% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{collect_data}
\alias{collect_data}
\title{Collect the data from an HDF5 file}
\usage{
collect_data(path = NULL)
}
\arguments{
\item{path}{character, path to the h5 file}
}
\value{
NA
}
\description{
This inserts data into the specified table
}
